requirements.txt
fastapi[all]
python-jose[cryptography]
passlib[bcrypt]
psycopg2-binary
requests